requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Image Processing
pillow==10.2.0

# Machine learning (MobileNetV3 — PyTorch)
torch==2.2.0
torchvision==0.17.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
